requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
langchain
langchain-openai
python-slugify
pyairtable